fix(mojo-0263): fix Int→Float64 in mixed_precision_simd + StepFn trait for training_loop
- test_mixed_precision_simd.mojo:225 — explicit Float64() cast around (i % 10)
- shared/training/script_runner.mojo — add StepFn trait + parametric overload
  of run_epoch_with_batches[S: StepFn] to work around nonescaping closures
- shared/training/__init__.mojo — export StepFn from training package
- test_training_loop.mojo:697 — migrate step_fn closure to @fieldwise_init
  SimpleStepFn struct implementing StepFn

+++ tests/shared/training/test_mixed_precision_simd.mojo
@@ -222,7 +222,7 @@ def test_roundtrip_fp16_fp32_fp16() raises:
     var fp16_ptr = fp16_original._data.bitcast[Float16]()
 
     for i in range(64):
-        fp16_ptr[i] = Float16(Float32((i % 10) / 10.0))
+        fp16_ptr[i] = Float16(Float32(Float64(i % 10) / 10.0))
 
     # Convert to FP32
     var fp32_intermediate = convert_to_fp32_master(fp16_original)
